refact: create lib/fraud
