+++ server/graphql/v1/mutations/orders.js
@@ -5,7 +5,6 @@ import config from 'config';
 import debugLib from 'debug';
 import * as hcaptcha from 'hcaptcha';
 import { get, isEmpty, isEqual, isNil, omit, pick, set } from 'lodash';
-import { isEmail } from 'validator';
 
 import activities from '../../../constants/activities';
 import CAPTCHA_PROVIDERS from '../../../constants/captcha-providers';
@@ -15,7 +14,8 @@ import status from '../../../constants/order_status';
 import { PAYMENT_METHOD_TYPE } from '../../../constants/paymentMethods';
 import roles from '../../../constants/roles';
 import { VAT_OPTIONS } from '../../../constants/vat';
-import cache, { purgeCacheForCollective } from '../../../lib/cache';
+import { purgeCacheForCollective } from '../../../lib/cache';
+import { checkGuestContribution, checkOrdersLimit, cleanOrdersLimit } from '../../../lib/fraud';
 import * as github from '../../../lib/github';
 import { getOrCreateGuestProfile } from '../../../lib/guest-accounts';
 import logger from '../../../lib/logger';
@@ -23,7 +23,7 @@ import * as libPayments from '../../../lib/payments';
 import recaptcha from '../../../lib/recaptcha';
 import { getChargeRetryCount, getNextChargeAndPeriodStartDates } from '../../../lib/recurring-contributions';
 import { canUseFeature } from '../../../lib/user-permissions';
-import { formatCurrency, md5, parseToBoolean, sleep } from '../../../lib/utils';
+import { formatCurrency, parseToBoolean } from '../../../lib/utils';
 import models from '../../../models';
 import { canRefund } from '../../common/transactions';
 import {
@@ -36,135 +36,11 @@ import {
   ValidationFailed,
 } from '../../errors';
 
-const oneHourInSeconds = 60 * 60;
-
 const debug = debugLib('orders');
 
 export const ORDER_PUBLIC_DATA_FIELDS = {
   pledgeCurrency: 'thegivingblock.pledgeCurrency',
   pledgeAmount: 'thegivingblock.pledgeAmount',
-};
-
-function getOrdersLimit(order, reqIp, reqMask) {
-  const limits = [];
-
-  const ordersLimits = config.limits.ordersPerHour;
-  const collectiveId = get(order, 'collective.id');
-  const fromCollectiveId = get(order, 'fromCollective.id');
-  const userEmail = get(order, 'user.email');
-  const guestInfo = get(order, 'guestInfo');
-
-  if (fromCollectiveId) {
-    // Limit on authenticated users
-    limits.push({
-      key: `order_limit_on_account_${fromCollectiveId}`,
-      value: ordersLimits.perAccount,
-    });
-    if (collectiveId) {
-      limits.push({
-        key: `order_limit_on_account_${fromCollectiveId}_and_collective_${collectiveId}`,
-        value: ordersLimits.perAccountForCollective,
-      });
-    }
-  } else {
-    // Limit on first time users
-    if (userEmail) {
-      const emailHash = md5(userEmail);
-      limits.push({
-        key: `order_limit_on_email_${emailHash}`,
-        value: ordersLimits.perEmail,
-      });
-      if (collectiveId) {
-        limits.push({
-          key: `order_limit_on_email_${emailHash}_and_collective_${collectiveId}`,
-          value: ordersLimits.perEmailForCollective,
-        });
-      }
-    }
-    // Limit on IPs
-    if (reqIp) {
-      limits.push({
-        key: `order_limit_on_ip_${md5(reqIp)}`,
-        value: ordersLimits.perIp,
-      });
-    }
-  }
-
-  if (reqMask && config.limits.enabledMasks.includes(reqMask)) {
-    limits.push({
-      key: `order_limit_on_mask_${reqMask}`,
-      value: ordersLimits.perMask,
-    });
-  }
-
-  // Guest Contributions
-  if (guestInfo && collectiveId) {
-    limits.push({
-      key: `order_limit_to_account_${collectiveId}`,
-      value: ordersLimits.forCollective,
-    });
-  }
-
-  return limits;
-}
-
-async function checkOrdersLimit(order, reqIp, reqMask) {
-  if (['ci', 'test'].includes(config.env)) {
-    return;
-  }
-
-  debug(`checkOrdersLimit reqIp:${reqIp} reqMask:${reqMask}`);
-
-  // Generic error message
-  // const errorMessage = 'Error while processing your request, please try again or contact support@opencollective.com.';
-  const errorMessage = 'Your card was declined.';
-
-  const limits = getOrdersLimit(order, reqIp, reqMask);
-
-  for (const limit of limits) {
-    const count = (await cache.get(limit.key)) || 0;
-    debug(`${count} orders for limit '${limit.key}'`);
-    const limitReached = count >= limit.value;
-    cache.set(limit.key, count + 1, oneHourInSeconds);
-    if (limitReached) {
-      debug(`Order limit reached for limit '${limit.key}'`);
-      // Slow down
-      await sleep(Math.random() * 1000 * 5);
-      // Show a developer-friendly message in DEV
-      if (config.env === 'development') {
-        throw new Error(`${errorMessage} Orders limit reached.`);
-      } else {
-        throw new Error(errorMessage);
-      }
-    }
-  }
-}
-
-async function cleanOrdersLimit(order, reqIp, reqMask) {
-  const limits = getOrdersLimit(order, reqIp, reqMask);
-
-  for (const limit of limits) {
-    cache.delete(limit.key);
-  }
-}
-
-const checkGuestContribution = async (order, loaders) => {
-  const { guestInfo } = order;
-
-  const collective = order.collective.id && (await loaders.Collective.byId.load(order.collective.id));
-  if (!collective) {
-    throw new BadRequest('Guest contributions need to be made to an existing collective');
-  }
-
-  if (!guestInfo) {
-    throw new BadRequest('You need to provide a guest profile with an email for logged out contributions');
-  } else if (!guestInfo.email || !isEmail(guestInfo.email)) {
-    throw new BadRequest('You need to provide a valid email');
-  } else if (order.fromCollective) {
-    throw new BadRequest('You need to be logged in to specify a contributing profile');
-  } else if (order.paymentMethod?.id || order.paymentMethod?.uuid) {
-    throw new BadRequest('You need to be logged in to be able to use an existing payment method');
-  }
 };
 
 const mustUpdateLocation = (existingLocation, newLocation) => {
